added one doc and renamed files to prevent confusion
