+++ packages/math-extras/src/rectangleExtras.ts
